requirements/requirements.txt
creme>=0.4.3
fbprophet>=0.5
matplotlib>=3.1.1
numpy>=1.17.3
pandas>=0.25.3
scipy>=1.3.1
scikit-learn>=0.21.3
tensorflow-gpu==2.0.0
tensorflow_probability>=0.8
